Implemented a extremely simple clickable button
@ this speed it will take me ages to implement anything remotely like an
UI x-D

It also feels a bit weird about how the signal stuff is working (I could
use the signals with arguments, but the C++ signal lib I was planning to
steal does not support that).

The textmetrics is a bit useless, how do I know the height of the
string? I can prolly estimate it for the most *normal* fonts (Here it
should be around 14px high, per default).